requirements.txt
flask==2.3.3
python-dotenv==1.0.0
google-generativeai==0.3.2
pandas==2.0.3
flask-cors==4.0.0
gunicorn==21.2.0  # For deployment